Make Ring Container resources configurable
This adds three properties managing three mid-frame containers. All of
them prefix with usual "prism.d3d12":
- .mainRingBufferThreshold - determines the size of Main Ring Buffer's
  flush threshold
- .constantRingBufferThreshold - determines the size of Constant Ring
  Buffer's flush threshold
- srvRingHeapThreshold - determines the size of CBV/SRV/UAV Ring
  Descriptor Heap's flush threshold

All three containers default to total size of 3 times the threshold.

This allows for changing the size of these containers without
recompiling the whole backend, which can be useful in optimizing the
backend on different hardware.

+++ modules/javafx.graphics/src/main/native-prism-d3d12/Internal/D3D12RingDescriptorHeap.cpp
@@ -65,7 +65,7 @@ bool RingDescriptorHeap::Init(D3D12_DESCRIPTOR_HEAP_TYPE type, bool shaderVisibl
     D3D12_DESCRIPTOR_HEAP_DESC desc;
     D3D12NI_ZERO_STRUCT(desc);
     desc.Type = type;
-    desc.NumDescriptors = mSize;
+    desc.NumDescriptors = static_cast<UINT>(mSize);
     desc.Flags = shaderVisible ? D3D12_DESCRIPTOR_HEAP_FLAG_SHADER_VISIBLE : D3D12_DESCRIPTOR_HEAP_FLAG_NONE;
     // TODO: D3D12: for multi-adapters, we need to set below to non-zero.
     // See: https://learn.microsoft.com/en-us/windows/win32/api/d3d12/ns-d3d12-d3d12_descriptor_heap_desc
